add additional output for look around mode, flips the look around angle, decreases the range of the intitial look around angles
