+++ etc/configuration/default.json
@@ -691,8 +691,8 @@
   },
   "look_around": {
     "look_around_timeout": {
-      "nanos": 0,
-      "secs": 1
+      "nanos": 800000000,
+      "secs": 0
     },
     "quick_search_timeout": {
       "nanos": 300000000,
@@ -703,19 +703,19 @@
       "pitch": 0.4
     },
     "left_positions": {
-      "yaw": -2.1,
+      "yaw": 1.3,
       "pitch": 0.0
     },
     "right_positions": {
-      "yaw": 2.1,
+      "yaw": -1.3,
       "pitch": 0.0
     },
     "halfway_left_positions": {
-      "yaw": -0.9,
+      "yaw": 0.9,
       "pitch": 0.0
     },
     "halfway_right_positions": {
-      "yaw": 0.9,
+      "yaw": -0.9,
       "pitch": 0.0
     }
   },
